C++17 support & third-party submodules

Text from Clinkscales's pull request:
----------------------------------------------------
<img width="150" height="146" alt="get-boost" src="https://github.com/user-attachments/assets/01083994-9329-43a7-9678-d6ce3214335c" />

Two changes in one PR, oh no. These ended up together because supporting C++17 meant adding new dependencies (Boost and fmt), which meant untangling how creaky dependency management has been.

First, on the C++17 side, mostly straightforward stuff:
- Replace calls to `std::format()` and `std::print()` with API-compatible, third-party `fmt` library (`fmt::format()`, `fmt::print()`).
- Remove use of `#elifdef`, converting to `#elif defined()`.
- Use template parameters where `auto&` parameters were used -- those are sugar for templates anyway.
- Hardcode the PI constant.
- Replace C++20 container functions (`contains()` -> `find() == end()`, `try_emplace()` -> `insert()`).
- Don't use container `emplace` with aggregate types.
- Use Boost `atomic` for floating-point atomic types.
- Replace `std::bit_cast()` with a local implementation, utilizing well-defined undefined behavior (`memcpy`).
- Convince the WinRT headers we're extracting a `com_ptr` from that we're not actually using any coroutine stuff.
- Remove designated initializers from the SDL event mapper. This one really hurts.
- Fix a bunch of new compiler warnings, mostly narrowing conversions.

---

Dependency management was a harder problem.

All third-party dependencies are now acquired as submodules. Helpfully, Boost allows acquiring specific modules as-needed, dramatically limiting the necessary footprint; SDL3_ttf likewise allows acquiring its dependencies through its own submodules. Mollybuild can take care of doing this automatically. That's the easy part, and it's already enough to integrate all dependencies into the Meson build, *except* for Boost, SDL3, and SDL3_ttf, which all need to be built from source. (Yes, Boost's header-only libraries need to be built in order to generate said headers.)

The hard part: Meson's idiomatic support for untangling this is utterly broken.

1. Meson would prefer we used its ghetto package management system, "Wrap" / "WrapDB", for dependencies. SDL3 and SDL3_ttf are provided, but not up-to-date. Even so, Boost is not provided at all; we'd have to manually write a porting shim. Even so, attempting to build SDL3 with Clang on Windows leads to mismanaged system includes, pulling in incompatible SSE-intrinsic headers. Even so, SDL3_ttf fails to correctly provide an import library for linking on Windows, either way.

2. Meson, alternatively, provides a cmake module for interoperation with CMake projects (which all of these dependencies provide as first- or second-class build options). However, this module doesn't allow propagating CMake inter-dependencies, meaning we can't pass the built SDL dependency to SDL3_ttf. Further, it mishandles special behavior in Boost's include-directory formatting on Windows.

Thus, we've got to do something manual.

My preference would be to write our own CMake shim, that could directly integrate into the Meson build, allowing fine-grain tracking of output files and etc., but this would effectively mean rewriting Meson's CMake module, which is in its direly-insufficient state for a reason.

Instead, work with the assumption that third-party dependencies are intended as a "one-time" setup problem, rather than something expected to require rebuilding as part of ordinary project workflow, and build them as part of the submodule initialization step. Then, introduce some helper scripts to obtain the resulting output info, so Meson can ingest it.

Meson's not very good at ingesting file data. We need to read from standard output, then use some brittle search-and-substring logic to extract the relevant information.

The new behavior is exposed by a new mollybuild command, `init`. This gets all submodules, all submodules of submodules, and then builds what needs building. This also makes it the better place to acquire pip dependencies, which now includes CMake. To disambiguate, formerly-`setup` has been renamed `build`, as it is actually a full build step that produces compiled output. (The previous name was translated from Meson's `setup`, which was configure-only; mollybuild has drifted significantly.)

This approach successfuly maintains the build's goal of providing quick, clean-room support for standing up a functioning build. There's only one new step (`mollybuild.py init`), for the benefit of removing several previous implicit steps of dependency acquisition. Even support for user-selected toolchains is maintained. However, it's not fully robust:
- Each dependency has a manual shim. They're easy to stand up, but new dependencies will always require new build scripts. This is unsolvable with the Meson + submodules scheme; the entire problem here is the auto-systems don't work.
- Upgrading or editing third-party code requires a full re-init of the project. We'd have to fully integrate the Meson build to resolve this.
- The core Meson build now depends on well-defined build paths. We could probably get around the latter issue by specifying a build directory to mollybuild, which could then propagate accordingly.

This is a draft because it's only been tested on Windows, with MSVC and Clang. I can formally request a merge once I've fixed whatever fallout is bound to happen from testing on Linux.

+++ src/colors.cpp
@@ -439,7 +439,7 @@ void ColorPoint::MutateChannels(glm::vec3 NewChannels)
 std::tuple<uint8_t, uint8_t, uint8_t> ColorPoint::To8BitRGB() const
 {
     const ColorPoint RGBColor = Encoding == ColorSpace::sRGB ? *this : Encode(ColorSpace::sRGB);
-    auto To8Bit = [](float Channel) -> uint8_t { return std::min(std::max(int(Channel * 255.0f), 0), 255); };
+    auto To8Bit = [](float Channel) -> uint8_t { return static_cast<uint8_t>(std::min(std::max(int(Channel * 255.0f), 0), 255)); };
 
     return
     {
