+++ src/sdl_time.cpp
@@ -14,7 +14,7 @@ namespace Time
             return;
         }
 
-        const uint64_t FramerateInterval = (1.0 / Framerate) * 1e+9;
+        const uint64_t FramerateInterval = static_cast<uint64_t>((1.0 / Framerate) * 1.0e+9);
         const uint64_t TimeSinceLastTick = SDL_GetTicksNS() - this->LastTick;
         const uint64_t Zero = 0;
         const uint64_t TimeToDelay = std::max(Zero, FramerateInterval - TimeSinceLastTick);
